Make fetch_control_units_contacts a shared task
